[HUGE] Fix makefile!
Due to not handling main.o properly, I got into nasty memory bugs that
led to segfaults. Symptom was that as soon as I added a vector to
GameState, I got weird nonintended behaviours, such as the GameState
screenWidth and Height not being set properly and the ships being placed
incorrectly. This happened even if I added a useless vector like a
vector<int>.